requirements.txt
setuptools >= 41.0.0
h5py >= 2.9.0
h5py-cache >= 1.0
requests >= 2.25.0
tqdm >= 4.43.0
pandas >= 1.1.4
Pillow >= 8.0.1
matplotlib >= 3.3.3

tensorboard >= 2.0
tensorboardx >= 2.0
opencv-python >= 4.2.0
scikit-learn >= 0.19.1
scikit-image >= 0.18.1
slackclient >= 2.5.0
reportlab >= 3.5.23
scipy >= 1.4.1
ImageHash >= 4.1.0
pytorch-lightning >= 1.2.2
torchsummary >= 1.5.1
wandb >= 0.10.10
faiss-gpu == 1.6.3
